+++ packages/client/src/index.ts
@@ -2,3 +2,4 @@ export * from './client';
 export * from './constants';
 export * from './signing';
 export * from './tx';
+export * from './types';