+++ packages/client/src/types/generics.ts
@@ -1,0 +1,5 @@
+export type InstanceTypeMap<T extends object> = {
+  [P in keyof T]: T[P] extends new (...args: any) => any
+    ? InstanceType<T[P]>
+    : T[P];
+};