+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.1.0](https://github.com/bitsongofficial/bitsongjs/compare/@bitsongjs/utils@2.1.0-beta.11...@bitsongjs/utils@2.1.0) (2022-09-14)
+
+**Note:** Version bump only for package @bitsongjs/utils
+
+
+
+
+
 # [2.1.0-beta.11](https://github.com/bitsongofficial/bitsongjs/compare/@bitsongjs/utils@2.1.0-beta.10...@bitsongjs/utils@2.1.0-beta.11) (2022-09-14)
 
 **Note:** Version bump only for package @bitsongjs/utils